Extract AzureProfile related code into profiles.py - to be shared with profiles creator tool
